Add initial version of advisor script.
Right now it prints the top 10 most profitable contracts
to invest in, but I'd ideally I'd like it to prioritize
contracts with an approaching end date. It could also
take in the amount that you want to bet, and tell you
exactly which contracts to divide up the bets among.